drafts/dockerized/requirements.txt
Django
djangorestframework
django-cors-headers
celery
psycopg2-binary
gunicorn
redis
twilio
requests
qrcode
pillow
pytz
selenium
webdriver_manager
